fix no session values
