--- src/app/components/providers/index.tsx
@@ -1,4 +1,0 @@
-import SessionProvider from './session'
-import {DynamicProvider} from './dynamic'
-
-export {SessionProvider, DynamicProvider}
